api/requirements.txt
flask==1.1.1
flask_cors==3.0.8
gunicorn==20.0.4
psycopg2-binary==2.8.4